% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_platforms}
\alias{add_platforms}
\alias{add_regions}
\alias{add_players}
\title{Join Platform/Region info to Runs}
\usage{
add_platforms(runs, platforms = speedrunr::platforms)

add_regions(runs, regions = speedrunr::regions)

add_players(runs)
}
\arguments{
\item{runs}{A \code{tbl} of runs as returned by \code{get_runs} or \code{get_leaderboards}}

\item{platforms, regions}{The platform/region data to use. Uses packaged datasets by default.}
}
\value{
The input \code{runs} tbl with resolved \code{system_}* variables.
}
\description{
Join Platform/Region info to Runs
}
\examples{
\dontrun{
runs <- get_leaderboards(game = "j1l9qz1g")

add_platforms(runs)
add_regions(runs)
}
}
