% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_games.R
\name{get_categories}
\alias{get_categories}
\title{Get A Game's Categories}
\source{
\url{https://github.com/speedruncomorg/api/blob/master/version1/games.md#get-gamesidcategories}
}
\usage{
get_categories(id, ...)
}
\arguments{
\item{id}{THe game's id as required by the API. See \link{get_games} to retrieve the id.}

\item{...}{Optional named parameters passed to the api}
}
\value{
A \link[tibble:tibble]{tibble::tibble} of categories
}
\description{
Get A Game's Categories
}
\examples{
\dontrun{
get_categories(id = "j1l9qz1g")
}
}
