% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game.R
\name{get_game}
\alias{get_game}
\title{Search for Games by Name}
\usage{
get_game(name = "sm64", ...)
}
\arguments{
\item{name}{A search query, \code{character}.}

\item{...}{Other named parameters passed to the API.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with search results
}
\description{
Search for Games by Name
}
\examples{
\dontrun{
get_game("Ocarina of Time")
}
}
