% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_leaderboard.R
\name{get_leaderboard}
\alias{get_leaderboard}
\title{Get A Leaderboard}
\usage{
get_leaderboard(game, category, level = NULL, top = 100, ...)
}
\arguments{
\item{game}{The game's \code{id} (mandatory).}

\item{category}{The category \code{id} (mandatory).}

\item{level}{The level \code{id}, if a per-level leaderboard is requested.}

\item{top}{\verb{[100]}. How many places to return. Note there might be multiple runs per place.}

\item{...}{Optional arguments to the API.}
}
\value{
A \link[tibble:tibble-package]{tibble} of length \code{top} with run data.
}
\description{
Get A Leaderboard
}
\examples{
# Get the Ocarina of Time 100\% leaderboard
\dontrun{
get_leaderboard(game = "j1l9qz1g", category = "q255jw2o")
}
}
