% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_platforms.R
\name{get_platforms}
\alias{get_platforms}
\title{Get All Platforms}
\source{
\url{https://github.com/speedruncomorg/api/blob/master/version1/platforms.md}
}
\usage{
get_platforms(max = 100, ...)
}
\arguments{
\item{max}{How many results to return. Default is \code{100}.}

\item{...}{Optional arguments passed to the API}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with platform info and ids.
}
\description{
Get All Platforms
}
\examples{
\dontrun{
get_platforms()
}
}
