% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regions.R
\name{get_regions}
\alias{get_regions}
\title{Get Regions}
\source{
\url{https://github.com/speedruncomorg/api/blob/master/version1/regions.md}
}
\usage{
get_regions(id = NULL, output = "df")
}
\arguments{
\item{id}{Optional \code{id} of a specific region, normally all regions are returned.}

\item{output}{Return a \code{df} (default) or named \code{list}.}
}
\value{
A \link[tibble:tibble-package]{tibble} of region \code{id} and \code{name}.
}
\description{
Get Regions
}
\examples{
\dontrun{
get_regions()
}
}
