% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_runs.R
\name{get_runs}
\alias{get_runs}
\title{Get All Runs Within a Game's Category}
\source{
\url{https://github.com/speedruncomorg/api/blob/master/version1/runs.md#get-runs}
}
\usage{
get_runs(game, category, max = 100, status = "verified", offset = NULL,
  verbose = FALSE, ...)
}
\arguments{
\item{game}{The game's id.}

\item{category}{The category's id.}

\item{max}{The number of runs to return, default is \code{100}.}

\item{status}{Filter by run status, default is \code{"verified"}. Leave blank for all runs.}

\item{offset}{Offset used for pagination, supply \code{200} to get the \emph{next} 200 runs.}

\item{verbose}{If \code{TRUE}, will display debugging output for pagination handling.
This option will probably not stick around for long.}

\item{...}{Optional arguments passed to API.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} of runs.
}
\description{
Get All Runs Within a Game's Category
}
\examples{
\dontrun{
# Get all Ocarina of Time 100\% runs:
runs <- get_runs(game = "j1l9qz1g", category = "q255jw2o")
}
}
