% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users.R
\name{get_user}
\alias{get_user}
\title{Get a Single User by ID}
\usage{
get_user(id, ...)
}
\arguments{
\item{id}{The user id, e.g. \code{"e8e5v680"} for \emph{zfg}.}

\item{...}{Optional parameters for the api.}
}
\value{
A \link[tibble:tibble-package]{tibble} with one row.
}
\description{
Get a Single User by ID
}
\examples{
\dontrun{
get_user(id = "e8e5v680")
}
}
