% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_variable}
\alias{get_variable}
\title{Get a Variable's Information}
\source{
\url{https://github.com/speedruncomorg/api/blob/master/version1/variables.md}
}
\usage{
get_variable(id, list_column = FALSE, ...)
}
\arguments{
\item{id}{The variable id.}

\item{list_column}{If \code{TRUE} (default) a list-column \code{values} is created.
If \code{FALSE}, esiting rows will be duplicated and merged with the \code{values} object,
yielding a flat tbl without nesting.}

\item{...}{Optional arguments passed to the API.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with the same number of rows as the variable has values if
\code{list_column} is \code{FALSE}, otherwise a single-row tbl with a list-column.
}
\description{
Get a Variable's Information
}
\examples{
\dontrun{
# Get Super Mario 64 variable for platform
get_variable(id = "e8m7em86")
}
}
