% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_games.R
\name{get_variables_game}
\alias{get_variables_game}
\title{Get a Game's Variables}
\source{
\url{https://github.com/speedruncomorg/api/blob/master/version1/games.md#get-gamesidvariables}
}
\usage{
get_variables_game(game, list_column = FALSE)
}
\arguments{
\item{game}{The game's \code{id}}

\item{list_column}{\verb{[FALSE]} Whether to return a list column or a flat \code{tbl}.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with one row per \code{variable} \emph{or}
per \code{value} depending on \code{list_column}
}
\description{
Get a Game's Variables
}
\examples{
# Get the variables for Ocarina of Time
\dontrun{
get_variables_game(game = "j1l9qz1g")
}
}
