% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_outlier}
\alias{is_outlier}
\title{Detect Outliers}
\usage{
is_outlier(x, method = "quantile", direction = "upper")
}
\arguments{
\item{x}{A \code{numeric} vector.}

\item{method}{Currently only \code{quantile} is supported.}

\item{direction}{\code{upper}, (default) \code{lower} or \code{both}.}
}
\value{
A \code{logical} vector of the same length as \code{x}
}
\description{
Detect Outliers
}
\examples{
is_outlier(runif(20))
}
