% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{platforms}
\alias{platforms}
\alias{regions}
\title{Platform and Region data from speedrun.com}
\format{\describe{
\item{platforms}{A \code{tibble} with 99 rows and 3 variables: \code{id, name, released}}
\item{regions}{A \code{tibble} with 6 rows and 2 variables: \code{id, name}}
}}
\usage{
platforms

regions
}
\description{
These datasets are saved for convenience, as they are relatively small and
not expected to change frequently.
}
\keyword{datasets}
