% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedrunr-package.R
\docType{package}
\name{speedrunr-package}
\alias{speedrunr}
\alias{speedrunr-package}
\title{speedrunr: Access Data from 'speedrun.com'}
\description{
Wrappers for the 'speedrun.com' API to make it easy
    to access run data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jemus42/speedrunr}
  \item Report bugs at \url{https://github.com/jemus42/speedrunr/issues}
}

}
\author{
\strong{Maintainer}: Lukas Burk \email{lukas@quantenbrot.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID})

}
\keyword{internal}
