% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{sr_get}
\alias{sr_get}
\title{Call the speedrun.com API}
\usage{
sr_get(path = "", ...)
}
\arguments{
\item{path}{API path.}

\item{...}{Optional parameters}
}
\value{
A \link{httr} response object
}
\description{
Call the speedrun.com API
}
\examples{
\dontrun{
# Get all users
sr_get(path = "users")
}
}
