% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_games.R
\name{get_games}
\alias{get_games}
\title{Search for Games by Name}
\source{
\url{https://github.com/speedruncomorg/api/blob/master/version1/games.md#get-games}
}
\usage{
get_games(name = "sm64", abbreviation = NULL, ...)
}
\arguments{
\item{name}{A search query, \code{character}.}

\item{abbreviation}{An exact abbreviation as listed on \href{https://speedrun.com}{speedrun.com}.
If this is set, \code{name} will be ignored.}

\item{...}{Other named parameters passed to the API.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with search results
}
\description{
Search for Games by Name
}
\examples{
\dontrun{
get_games("Ocarina of Time")
}
}
