% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_platforms.R
\name{get_platforms}
\alias{get_platforms}
\title{Get Platforms}
\source{
\url{https://github.com/speedruncomorg/api/blob/master/version1/platforms.md}
}
\usage{
get_platforms(id = NULL, output = "df")
}
\arguments{
\item{id}{Optional \code{id} of a specific platform, normally all platforms are returned.}

\item{output}{Return a \code{df} (default) or named \code{list}.}
}
\value{
A \link[tibble:tibble-package]{tibble} with platform info and ids.
}
\description{
Get Platforms
}
\note{
Currently the \code{released} variables is dropped on \code{list} output because I am bad at lists.
}
\examples{
\dontrun{
get_platforms()
}
}
